// frontend/src/definitions/interfaces.ts
export interface data {
    bgcolor: string;
    fontcolor: string;
    inputbgcolor: string;
    btncolor: string;
    btntxtcolor: string;
    inputtxtcolor: string;
    name: string;
    email: string;
    title: string;
    reply_email: boolean;
    reply_email_content: string;
    namefield: boolean;
    emailfield: boolean;
    bodyfield: boolean;
    status: boolean;
    premium: boolean;
  }